refactor: post service에서 category typeorm repo -> prisma
